Game class refactored to add deck class.
