+++ src/main/java/models/Card.java
@@ -5,13 +5,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.io.Serializable;
 
-/**
- * Assignment 1: No changes are needed in this file, but it is good to be aware of for future assignments.
- */
-
 public class Card implements Serializable {
     public final int value;
     public final Suit suit;
+
+    public java.util.List<java.util.List<Card>> cols = new ArrayList<>(4);
 
     @JsonCreator
     public Card(@JsonProperty("value") int value, @JsonProperty("suit") Suit suit) {
